requirements.txt
fastapi
uvicorn
numpy
python-dotenv
ffmpeg-python
torch
transformers
huggingface_hub
bitsandbytes
python-multipart
aiohttp
websockets
pydub
# if using faster-whisper
faster-whisper
# if using xtts
coqui-tts==0.26.0
# if using deepspeed in xtts
deepspeed
accelerate
# if using vllm implementations
vllm==0.6.4

# if using sensevoice stt
#funasr

# if using whisper-jax implementation
# this is a fork of whisper-jax with a dependency issue fix
# install separately afer requirements.txt
#git+https://github.com/naman14/whisper-jax.git

#apple silicon
# mlx-lm
# lightning-whisper-mlx

#==optional==
#cartesia
#elevenlabs

# For Opus codec support
sphn
scipy
